requirements.txt
python-nmap
netifaces
colorama
tqdm
prettytable